services/self-healing-agent/requirements.txt
docker==7.0.0
psycopg2-binary==2.9.9
requests==2.31.0
psutil==5.9.6
python-dotenv==1.0.0
